+++ example/cli/src/app/app.component.ts
@@ -5,6 +5,7 @@ import { Component } from '@angular/core';
   template: `
     <ul>
       <li><a routerLink="/basic">Basic</a></li>
+      <li><a routerLink="/templates">Templates</a></li>
     </ul>
     <router-outlet></router-outlet>
   `