+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Fixed tree-node and tree-node-level-X classes.
 * Removed redundant div in children
 * Added integration tests
+* Passing node in context to loading template
 
 <a name="3.2.3"></a>
 # 3.2.3 (2017-03-08)